javis-langchain/app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
deprecation==2.1.0
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
grpcio==1.73.1
grpcio-health-checking==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
packaging==25.0
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
tokenizers==0.19.1
torch==2.7.1
tqdm==4.67.1
transformers==4.41.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
weaviate-client==3.26.7
